requirements.txt
pyzmq>=25.0.0
pyyaml>=6.0
loguru>=0.7.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
requests>=2.31.0
websocket-client>=1.6.0
